validation,migration

+++ database/migrations/2019_08_25_084455_create_internet_offers_table.php
@@ -16,11 +16,11 @@ class CreateInternetOffersTable extends Migration
         Schema::create('internet_offers', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('title');
-            $table->integer('volume');
+            $table->bigInteger('volume');
             $table->string('validity');
-            $table->integer('price')->unsigned();
+            $table->bigInteger('price')->unsigned();
             $table->text('offer_code');
-            $table->integer('points')->unsigned();
+            $table->bigInteger('points')->unsigned();
             $table->timestamps();
         });
     }
